+++ cloudant-sync-datastore-core/src/test/java/com/cloudant/common/AdvancedAPITest.java
@@ -1,0 +1,20 @@
+package com.cloudant.common;
+
+import com.cloudant.sync.documentstore.advanced.Database;
+
+import org.junit.Before;
+
+/**
+ * Superclass that exposes the "advanced" Database APIs for testing
+ */
+public class AdvancedAPITest extends DocumentStoreTestBase {
+
+    protected Database advancedDatabase;
+
+    @Before
+    public void setUpAdvanced() throws Exception {
+        advancedDatabase = documentStore.advanced();
+    }
+
+
+}